move file source & tree preview data fetching to server
