+++ packages/web/src/app/[domain]/browse/[...path]/components/codePreviewPanel.tsx
@@ -1,6 +1,4 @@
-'use client';
-
-import { getCodeHostInfoForRepo, unwrapServiceError } from "@/lib/utils";
+import { getCodeHostInfoForRepo, isServiceError, unwrapServiceError } from "@/lib/utils";
 import { useBrowseParams } from "@/app/[domain]/browse/hooks/useBrowseParams";
 import { useQuery } from "@tanstack/react-query";
 import { getFileSource } from "@/features/search/fileSourceApi";
@@ -10,53 +8,35 @@ import { Separator } from "@/components/ui/separator";
 import { getRepoInfoByName } from "@/actions";
 import { cn } from "@/lib/utils";
 import Image from "next/image";
-import { useMemo } from "react";
 import { PureCodePreviewPanel } from "./pureCodePreviewPanel";
 import { PathHeader } from "@/app/[domain]/components/pathHeader";
 
-export const CodePreviewPanel = () => {
-    const { path, repoName, revisionName } = useBrowseParams();
-    const domain = useDomain();
+interface CodePreviewPanelProps {
+    path: string;
+    repoName: string;
+    revisionName?: string;
+    domain: string;
+}
 
-    const { data: fileSourceResponse, isPending: isFileSourcePending, isError: isFileSourceError } = useQuery({
-        queryKey: ['fileSource', repoName, revisionName, path, domain],
-        queryFn: () => unwrapServiceError(getFileSource({
-            fileName: path,
-            repository: repoName,
-            branch: revisionName
-        }, domain)),
-    });
+export const CodePreviewPanel = async ({ path, repoName, revisionName, domain }: CodePreviewPanelProps) => {
+    const fileSourceResponse = await getFileSource({
+        fileName: path,
+        repository: repoName,
+        branch: revisionName,
+    }, domain);
 
-    const { data: repoInfoResponse, isPending: isRepoInfoPending, isError: isRepoInfoError } = useQuery({
-        queryKey: ['repoInfo', repoName, domain],
-        queryFn: () => unwrapServiceError(getRepoInfoByName(repoName, domain)),
-    });
+    const repoInfoResponse = await getRepoInfoByName(repoName, domain);
 
-    const codeHostInfo = useMemo(() => {
-        if (!repoInfoResponse) {
-            return undefined;
-        }
-
-        return getCodeHostInfoForRepo({
-            codeHostType: repoInfoResponse.codeHostType,
-            name: repoInfoResponse.name,
-            displayName: repoInfoResponse.displayName,
-            webUrl: repoInfoResponse.webUrl,
-        });
-    }, [repoInfoResponse]);
-
-    if (isFileSourcePending || isRepoInfoPending) {
-        return (
-            <div className="flex flex-col w-full min-h-full items-center justify-center">
-                <Loader2 className="w-4 h-4 animate-spin" />
-                Loading...
-            </div>
-        )
-    }
-
-    if (isFileSourceError || isRepoInfoError) {
+    if (isServiceError(fileSourceResponse) || isServiceError(repoInfoResponse)) {
         return <div>Error loading file source</div>
     }
+
+    const codeHostInfo = getCodeHostInfoForRepo({
+        codeHostType: repoInfoResponse.codeHostType,
+        name: repoInfoResponse.name,
+        displayName: repoInfoResponse.displayName,
+        webUrl: repoInfoResponse.webUrl,
+    });
 
     return (
         <>
